+++ packages/react-form-nextjs/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @tanstack/react-form-nextjs
 
+## 1.27.2
+
+### Patch Changes
+
+- use React 18's useId hook by default for formId generation, only calling Math.random() as a fallback if no formId is provided. ([#1913](https://github.com/TanStack/form/pull/1913))
+
+- Updated dependencies [[`6a5e1c1`](https://github.com/TanStack/form/commit/6a5e1c1a4f0f0519705cc5fd15cbe8afb878a42d), [`9ce5f28`](https://github.com/TanStack/form/commit/9ce5f286a12fb0f1662f3eeeff14c75d8f3e4417)]:
+  - @tanstack/react-form@1.27.2
+
+## 1.27.1
+
+### Patch Changes
+
+- Updated dependencies [[`3b080ec`](https://github.com/TanStack/form/commit/3b080ec1faefa9894c0f73880dbff680888e6a9a)]:
+  - @tanstack/react-form@1.27.1
+
+## 1.27.0
+
+### Patch Changes
+
+- Updated dependencies [[`03c2bee`](https://github.com/TanStack/form/commit/03c2beed867f097ac61fb6411ce9cd5a9f3b4c58), [`8afbfc3`](https://github.com/TanStack/form/commit/8afbfc39d7373ec2b516f7c8ff5585ca44098cc1), [`4cd4068`](https://github.com/TanStack/form/commit/4cd4068014ee902399da6978becd443068b0c672)]:
+  - @tanstack/react-form@1.27.0
+
+## 1.26.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-form@1.26.0
+
 ## 1.25.0
 
 ### Patch Changes